move files into integration folder, move functions into helper file
